chore: clean up test setup and make it run in eclipse out-of-box
